% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_survey_data.R
\name{delete_survey_data}
\alias{delete_survey_data}
\title{Delete SURVEY Image Filenames}
\usage{
delete_survey_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "SURVEY"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the SURVEY images
}
