% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey_data.R
\name{download_survey_data}
\alias{download_survey_data}
\title{Download SURVEY Image Filenames}
\usage{
download_survey_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "SURVEY"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the SURVEY images
}
