% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_filenames.R
\name{get_survey_filenames}
\alias{get_survey_filenames}
\title{Get SURVEY Image Filenames}
\usage{
get_survey_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "SURVEY"} so it cannot be specified}
}
\value{
Vector of filenames
}
\description{
Return the filenames for the SURVEY images
}
\examples{
get_survey_filenames()
}
